feat: add comprehensive integration tests for GET endpoints without parameters
